pesado y excedente

+++ src/unq/edu/po2/servicio/Electricidad.java
@@ -13,12 +13,11 @@ public class Electricidad implements Servicio {
 		this.precioPorKw = precioPorKw;
 	}
 	
-	//Duration es de tipo long pero no creo que supere los 2.147 millones de horas (245 años)
 	@Override
 	public double getMontoFinal(Container container) {
 		ContainerReefer reefer = (ContainerReefer) container;
 		
-		int horas = (int) Duration.between(reefer.getMomentoConexion(), reefer.getMomentoDesconexion()).toHours();
+		long horas = Duration.between(reefer.getMomentoConexion(), reefer.getMomentoDesconexion()).toHours();
 
 		return horas * reefer.getConsumoXHora() * precioPorKw;
 	}
